@@ -1,20 +1,18 @@
 module github.com/wangluozhe/chttp
 
-go 1.20
+go 1.23.4
 
 require (
-	github.com/refraction-networking/utls v1.5.4
-	golang.org/x/net v0.17.0
-	golang.org/x/term v0.13.0
+	github.com/refraction-networking/utls v1.6.7
+	golang.org/x/net v0.23.0
+	golang.org/x/term v0.18.0
 )
 
 require (
-	github.com/andybalholm/brotli v1.0.5 // indirect
-	github.com/cloudflare/circl v1.3.3 // indirect
-	github.com/gaukas/godicttls v0.0.4 // indirect
-	github.com/klauspost/compress v1.16.7 // indirect
-	github.com/quic-go/quic-go v0.37.4 // indirect
-	golang.org/x/crypto v0.14.0 // indirect
-	golang.org/x/sys v0.13.0 // indirect
-	golang.org/x/text v0.13.0 // indirect
+	github.com/andybalholm/brotli v1.0.6 // indirect
+	github.com/cloudflare/circl v1.3.7 // indirect
+	github.com/klauspost/compress v1.17.4 // indirect
+	golang.org/x/crypto v0.21.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
 )
