@@ -5,19 +5,19 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/wangluozhe/chttp"
-	"github.com/wangluozhe/chttp/http2"
 	"io"
+	"net"
 	"net/url"
 	"strings"
 )
 
-var settings = map[string]http2.SettingID{
-	"HEADER_TABLE_SIZE":      http2.SettingHeaderTableSize,
-	"ENABLE_PUSH":            http2.SettingEnablePush,
-	"MAX_CONCURRENT_STREAMS": http2.SettingMaxConcurrentStreams,
-	"INITIAL_WINDOW_SIZE":    http2.SettingInitialWindowSize,
-	"MAX_FRAME_SIZE":         http2.SettingMaxFrameSize,
-	"MAX_HEADER_LIST_SIZE":   http2.SettingMaxHeaderListSize,
+var settings = map[string]http.HTTP2SettingID{
+	"HEADER_TABLE_SIZE":      http.HTTP2SettingHeaderTableSize,
+	"ENABLE_PUSH":            http.HTTP2SettingEnablePush,
+	"MAX_CONCURRENT_STREAMS": http.HTTP2SettingMaxConcurrentStreams,
+	"INITIAL_WINDOW_SIZE":    http.HTTP2SettingInitialWindowSize,
+	"MAX_FRAME_SIZE":         http.HTTP2SettingMaxFrameSize,
+	"MAX_HEADER_LIST_SIZE":   http.HTTP2SettingMaxHeaderListSize,
 }
 
 type H2Settings struct {
@@ -40,11 +40,11 @@ type H2Settings struct {
 	PriorityFrames []map[string]interface{} `json:"PriorityFrames"`
 }
 
-func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http2.HTTP2Settings) {
-	http2Settings = &http2.HTTP2Settings{
+func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http.HTTP2Settings) {
+	http2Settings = &http.HTTP2Settings{
 		Settings:       nil,
 		ConnectionFlow: 0,
-		HeaderPriority: &http2.PriorityParam{},
+		HeaderPriority: &http.HTTP2PriorityParam{},
 		PriorityFrames: nil,
 	}
 	if h2Settings.Settings != nil {
@@ -52,7 +52,7 @@ func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http2.HTTP2Settings
 			for _, orderKey := range h2Settings.SettingsOrder {
 				val := h2Settings.Settings[orderKey]
 				if val != 0 || orderKey == "ENABLE_PUSH" {
-					http2Settings.Settings = append(http2Settings.Settings, http2.Setting{
+					http2Settings.Settings = append(http2Settings.Settings, http.HTTP2Setting{
 						ID:  settings[orderKey],
 						Val: uint32(val),
 					})
@@ -60,7 +60,7 @@ func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http2.HTTP2Settings
 			}
 		} else {
 			for id, val := range h2Settings.Settings {
-				http2Settings.Settings = append(http2Settings.Settings, http2.Setting{
+				http2Settings.Settings = append(http2Settings.Settings, http.HTTP2Setting{
 					ID:  settings[id],
 					Val: uint32(val),
 				})
@@ -87,14 +87,14 @@ func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http2.HTTP2Settings
 		case float64:
 			streamDep = int(s.(float64))
 		}
-		var priorityParam *http2.PriorityParam
+		var priorityParam *http.HTTP2PriorityParam
 		if w == nil {
-			priorityParam = &http2.PriorityParam{
+			priorityParam = &http.HTTP2PriorityParam{
 				StreamDep: uint32(streamDep),
 				Exclusive: h2Settings.HeaderPriority["exclusive"].(bool),
 			}
 		} else {
-			priorityParam = &http2.PriorityParam{
+			priorityParam = &http.HTTP2PriorityParam{
 				StreamDep: uint32(streamDep),
 				Exclusive: h2Settings.HeaderPriority["exclusive"].(bool),
 				Weight:    uint8(weight - 1),
@@ -129,24 +129,24 @@ func ToHTTP2Settings(h2Settings *H2Settings) (http2Settings *http2.HTTP2Settings
 			case float64:
 				streamID = int(sid.(float64))
 			}
-			var priorityParam http2.PriorityParam
+			var priorityParam http.HTTP2PriorityParam
 			if w == nil {
-				priorityParam = http2.PriorityParam{
+				priorityParam = http.HTTP2PriorityParam{
 					StreamDep: uint32(streamDep),
 					Exclusive: priorityParamSource["exclusive"].(bool),
 				}
 			} else {
-				priorityParam = http2.PriorityParam{
+				priorityParam = http.HTTP2PriorityParam{
 					StreamDep: uint32(streamDep),
 					Exclusive: priorityParamSource["exclusive"].(bool),
 					Weight:    uint8(weight - 1),
 				}
 			}
-			http2Settings.PriorityFrames = append(http2Settings.PriorityFrames, http2.PriorityFrame{
-				FrameHeader: http2.FrameHeader{
+			http2Settings.PriorityFrames = append(http2Settings.PriorityFrames, http.HTTP2PriorityFrame{
+				HTTP2FrameHeader: http.HTTP2FrameHeader{
 					StreamID: uint32(streamID),
 				},
-				PriorityParam: priorityParam,
+				HTTP2PriorityParam: priorityParam,
 			})
 		}
 	}
@@ -185,9 +185,24 @@ func request(req *http.Request) {
 		},
 	}
 	h2ss := ToHTTP2Settings(h2s)
-	client := http.Client{Transport: &http2.Transport{HTTP2Settings: h2ss}}
-	resp, _ := client.Do(req)
-	text, _ := io.ReadAll(resp.Body)
+	t1 := &http.Transport{}
+	t2, err := http.HTTP2ConfigureTransports(t1)
+	if err != nil {
+		fmt.Println(err)
+	}
+	t2.HTTP2Settings = h2ss
+	t1.H2Transport = t2
+	proxyURL, _ := url.Parse("http://127.0.0.1:7890")
+	t1.Proxy = http.ProxyURL(proxyURL)
+	client := http.Client{Transport: t1}
+	resp, err := client.Do(req)
+	if err != nil {
+		fmt.Println(err)
+	}
+	text, err := io.ReadAll(resp.Body)
+	if err != nil {
+		fmt.Println(err)
+	}
 	fmt.Println(string(text))
 }
 
@@ -229,7 +244,7 @@ func get() {
 }
 
 func post() {
-	rawurl := "https://httpbin.org/post"
+	rawurl := "https://tls.peet.ws/api/all"
 	data := url.Values{}
 	data.Set("username", "example_user")
 	data.Set("password", "example_password")
@@ -270,7 +285,7 @@ func post() {
 }
 
 func post_json() {
-	rawurl := "https://httpbin.org/post"
+	rawurl := "https://tls.peet.ws/api/all"
 	json_data := map[string]interface{}{
 		"page":  "1",
 		"limit": 10,
