@@ -5,6 +5,8 @@
 package http
 
 import (
+	"github.com/wangluozhe/chttp/httptrace"
+	"github.com/wangluozhe/chttp/internal/ascii"
 	"io"
 	"net/textproto"
 	"sort"
@@ -12,13 +14,13 @@ import (
 	"sync"
 	"time"
 
-	"github.com/wangluozhe/chttp/httptrace"
+	"golang.org/x/net/http/httpguts"
 )
 
-// A Header represents the Key-value pairs in an HTTP header.
+// A Header represents the key-value pairs in an HTTP header.
 //
 // The keys should be in canonical form, as returned by
-// CanonicalHeaderKey.
+// [CanonicalHeaderKey].
 type Header map[string][]string
 
 // HeaderOrderKey is a magic Key for ResponseWriter.Header map keys
@@ -55,42 +57,53 @@ func (h Header) inUnChangedHeaderKeys(key string) string {
 	return key
 }
 
-// Add adds the Key, value pair to the header.
-// It appends to any existing Values associated with Key.
-// The Key is case insensitive; it is canonicalized by
-// CanonicalHeaderKey.
+// 判断字符串是否在UnChangedHeaderKeys中
+func (h Header) ContainsUnChangedHeaderKeys(target string) bool {
+	set := make(map[string]struct{}, len(h[UnChangedHeaderKey]))
+	for _, item := range h[UnChangedHeaderKey] {
+		set[item] = struct{}{}
+	}
+	_, exists := set[target]
+	return exists
+}
+
+// Add adds the key, value pair to the header.
+// It appends to any existing values associated with key.
+// The key is case insensitive; it is canonicalized by
+// [CanonicalHeaderKey].
 func (h Header) Add(key, value string) {
 	textproto.MIMEHeader(h).Add(key, value)
 }
 
-// Set sets the header entries associated with Key to the
-// single element value. It replaces any existing Values
-// associated with Key. The Key is case insensitive; it is
-// canonicalized by textproto.CanonicalMIMEHeaderKey.
+// Set sets the header entries associated with key to the
+// single element value. It replaces any existing values
+// associated with key. The key is case insensitive; it is
+// canonicalized by [textproto.CanonicalMIMEHeaderKey].
 // To use non-canonical keys, assign to the map directly.
 func (h Header) Set(key, value string) {
 	textproto.MIMEHeader(h).Set(key, value)
 }
 
-// Get gets the first value associated with the given Key. If
-// there are no Values associated with the Key, Get returns "".
-// It is case insensitive; textproto.CanonicalMIMEHeaderKey is
-// used to canonicalize the provided Key. To use non-canonical keys,
+// Get gets the first value associated with the given key. If
+// there are no values associated with the key, Get returns "".
+// It is case insensitive; [textproto.CanonicalMIMEHeaderKey] is
+// used to canonicalize the provided key. Get assumes that all
+// keys are stored in canonical form. To use non-canonical keys,
 // access the map directly.
 func (h Header) Get(key string) string {
 	return textproto.MIMEHeader(h).Get(key)
 }
 
-// Values returns all Values associated with the given Key.
-// It is case insensitive; textproto.CanonicalMIMEHeaderKey is
-// used to canonicalize the provided Key. To use non-canonical
+// Values returns all values associated with the given key.
+// It is case insensitive; [textproto.CanonicalMIMEHeaderKey] is
+// used to canonicalize the provided key. To use non-canonical
 // keys, access the map directly.
 // The returned slice is not a copy.
 func (h Header) Values(key string) []string {
 	return textproto.MIMEHeader(h).Values(key)
 }
 
-// get is like Get, but Key must already be in CanonicalHeaderKey form.
+// get is like Get, but key must already be in CanonicalHeaderKey form.
 func (h Header) get(key string) string {
 	if v := h[key]; len(v) > 0 {
 		return v[0]
@@ -98,16 +111,16 @@ func (h Header) get(key string) string {
 	return ""
 }
 
-// has reports whether h has the provided Key defined, even if it's
+// has reports whether h has the provided key defined, even if it's
 // set to 0-length slice.
 func (h Header) has(key string) bool {
 	_, ok := h[key]
 	return ok
 }
 
-// Del deletes the Values associated with Key.
-// The Key is case insensitive; it is canonicalized by
-// CanonicalHeaderKey.
+// Del deletes the values associated with key.
+// The key is case insensitive; it is canonicalized by
+// [CanonicalHeaderKey].
 func (h Header) Del(key string) {
 	textproto.MIMEHeader(h).Del(key)
 }
@@ -127,14 +140,20 @@ func (h Header) Clone() Header {
 		return nil
 	}
 
-	// Find total number of Values.
+	// Find total number of values.
 	nv := 0
 	for _, vv := range h {
 		nv += len(vv)
 	}
-	sv := make([]string, nv) // shared backing array for headers' Values
+	sv := make([]string, nv) // shared backing array for headers' values
 	h2 := make(Header, len(h))
 	for k, vv := range h {
+		if vv == nil {
+			// Preserve nil values. ReverseProxy distinguishes
+			// between nil and zero-length header values.
+			h2[k] = nil
+			continue
+		}
 		n := copy(sv, vv)
 		h2[k] = sv[:n:n]
 		sv = sv[n:]
@@ -150,7 +169,7 @@ var timeFormats = []string{
 
 // ParseTime parses a time header (such as the Date: header),
 // trying each of the three formats allowed by HTTP/1.1:
-// TimeFormat, time.RFC850, and time.ANSIC.
+// [TimeFormat], [time.RFC850], and [time.ANSIC].
 func ParseTime(text string) (t time.Time, err error) {
 	for _, layout := range timeFormats {
 		t, err = time.Parse(layout, text)
@@ -172,17 +191,14 @@ func (w stringWriter) WriteString(s string) (n int, err error) {
 	return w.w.Write([]byte(s))
 }
 
-type HeaderKeyValues struct {
-	Key    string
-	Values []string
+type keyValues struct {
+	key    string
+	values []string
 }
 
-// A headerSorter implements sort.Interface by sorting a []keyValues
-// by the given order, if not nil, or by Key otherwise.
-// It's used as a pointer, so it can fit in a sort.Interface
-// interface value without allocation.
+// headerSorter contains a slice of keyValues sorted by keyValues.key.
 type headerSorter struct {
-	kvs   []HeaderKeyValues
+	kvs   []keyValues
 	order map[string]int
 }
 
@@ -191,14 +207,14 @@ func (s *headerSorter) Swap(i, j int) { s.kvs[i], s.kvs[j] = s.kvs[j], s.kvs[i] 
 func (s *headerSorter) Less(i, j int) bool {
 	// If the order isn't defined, sort lexicographically.
 	if s.order == nil {
-		return s.kvs[i].Key < s.kvs[j].Key
+		return s.kvs[i].key < s.kvs[j].key
 	}
 	//idxi, iok := s.order[s.kvs[i].Key]
 	//idxj, jok := s.order[s.kvs[j].Key]
-	idxi, iok := s.order[strings.ToLower(s.kvs[i].Key)]
-	idxj, jok := s.order[strings.ToLower(s.kvs[j].Key)]
+	idxi, iok := s.order[strings.ToLower(s.kvs[i].key)]
+	idxj, jok := s.order[strings.ToLower(s.kvs[j].key)]
 	if !iok && !jok {
-		return s.kvs[i].Key < s.kvs[j].Key
+		return s.kvs[i].key < s.kvs[j].key
 	} else if !iok && jok {
 		return false
 	} else if iok && !jok {
@@ -211,51 +227,44 @@ var headerSorterPool = sync.Pool{
 	New: func() interface{} { return new(headerSorter) },
 }
 
-var mutex = &sync.RWMutex{}
-
-// SortedKeyValues returns h's keys sorted in the returned kvs
+// sortedKeyValues returns h's keys sorted in the returned kvs
 // slice. The headerSorter used to sort is also returned, for possible
 // return to headerSorterCache.
-func (h Header) SortedKeyValues(exclude map[string]bool) (kvs []HeaderKeyValues, hs *headerSorter) {
+func (h Header) sortedKeyValues(exclude map[string]bool) (kvs []keyValues, hs *headerSorter) {
 	hs = headerSorterPool.Get().(*headerSorter)
 	if cap(hs.kvs) < len(h) {
-		hs.kvs = make([]HeaderKeyValues, 0, len(h))
+		hs.kvs = make([]keyValues, 0, len(h))
 	}
 	kvs = hs.kvs[:0]
 	for k, vv := range h {
-		mutex.RLock()
 		if !exclude[k] {
-			kvs = append(kvs, HeaderKeyValues{k, vv})
+			kvs = append(kvs, keyValues{k, vv})
 		}
-		mutex.RUnlock()
 	}
 	hs.kvs = kvs
 	sort.Sort(hs)
 	return kvs, hs
 }
 
-func (h Header) SortedKeyValuesBy(order map[string]int, exclude map[string]bool) (kvs []HeaderKeyValues, hs *headerSorter) {
+func (h Header) sortedKeyValuesBy(order map[string]int, exclude map[string]bool) (kvs []keyValues, hs *headerSorter) {
 	hs = headerSorterPool.Get().(*headerSorter)
 	if cap(hs.kvs) < len(h) {
-		hs.kvs = make([]HeaderKeyValues, 0, len(h))
+		hs.kvs = make([]keyValues, 0, len(h))
 	}
 	kvs = hs.kvs[:0]
 	for k, vv := range h {
-		mutex.RLock()
 		if !exclude[k] {
-			kvs = append(kvs, HeaderKeyValues{k, vv})
+			kvs = append(kvs, keyValues{k, vv})
 		}
-		mutex.RUnlock()
 	}
 	hs.kvs = kvs
 	hs.order = order
 	sort.Sort(hs)
-
 	return kvs, hs
 }
 
 // WriteSubset writes a header in wire format.
-// If exclude is not nil, keys where exclude[Key] == true are not written.
+// If exclude is not nil, keys where exclude[key] == true are not written.
 // Keys are not canonicalized before checking the exclude map.
 func (h Header) WriteSubset(w io.Writer, exclude map[string]bool) error {
 	return h.writeSubset(w, exclude, nil)
@@ -267,14 +276,12 @@ func (h Header) writeSubset(w io.Writer, exclude map[string]bool, trace *httptra
 		ws = stringWriter{w}
 	}
 
-	var kvs []HeaderKeyValues
+	var kvs []keyValues
 	var sorter *headerSorter
 
-	mutex.Lock()
 	exclude[HeaderOrderKey] = true
 	exclude[PHeaderOrderKey] = true
 	exclude[UnChangedHeaderKey] = true
-	mutex.Unlock()
 
 	// Check if the HeaderOrder is defined.
 	if headerOrder, ok := h[HeaderOrderKey]; ok {
@@ -285,17 +292,24 @@ func (h Header) writeSubset(w io.Writer, exclude map[string]bool, trace *httptra
 		if exclude == nil {
 			exclude = make(map[string]bool)
 		}
-		kvs, sorter = h.SortedKeyValuesBy(order, exclude)
+		kvs, sorter = h.sortedKeyValuesBy(order, exclude)
 	} else {
-		kvs, sorter = h.SortedKeyValues(exclude)
+		kvs, sorter = h.sortedKeyValues(exclude)
 	}
 
 	var formattedVals []string
 	for _, kv := range kvs {
-		for _, v := range kv.Values {
+		if !httpguts.ValidHeaderFieldName(kv.key) {
+			// This could be an error. In the common case of
+			// writing response headers, however, we have no good
+			// way to provide the error back to the server
+			// handler, so just drop invalid headers instead.
+			continue
+		}
+		for _, v := range kv.values {
 			v = headerNewlineToSpace.Replace(v)
 			v = textproto.TrimString(v)
-			for _, s := range []string{kv.Key, ": ", v, "\r\n"} {
+			for _, s := range []string{kv.key, ": ", v, "\r\n"} {
 				s = h.inUnChangedHeaderKeys(s)
 				if _, err := ws.WriteString(s); err != nil {
 					headerSorterPool.Put(sorter)
@@ -307,7 +321,7 @@ func (h Header) writeSubset(w io.Writer, exclude map[string]bool, trace *httptra
 			}
 		}
 		if trace != nil && trace.WroteHeaderField != nil {
-			trace.WroteHeaderField(kv.Key, formattedVals)
+			trace.WroteHeaderField(kv.key, formattedVals)
 			formattedVals = nil
 		}
 	}
@@ -316,10 +330,10 @@ func (h Header) writeSubset(w io.Writer, exclude map[string]bool, trace *httptra
 }
 
 // CanonicalHeaderKey returns the canonical format of the
-// header Key s. The canonicalization converts the first
+// header key s. The canonicalization converts the first
 // letter and any letter following a hyphen to upper case;
 // the rest are converted to lowercase. For example, the
-// canonical Key for "accept-encoding" is "Accept-Encoding".
+// canonical key for "accept-encoding" is "Accept-Encoding".
 // If s contains a space or invalid header field bytes, it is
 // returned without modifications.
 func CanonicalHeaderKey(s string) string { return textproto.CanonicalMIMEHeaderKey(s) }
@@ -353,7 +367,7 @@ func hasToken(v, token string) bool {
 		if endPos := sp + len(token); endPos != len(v) && !isTokenBoundary(v[endPos]) {
 			continue
 		}
-		if strings.EqualFold(v[sp:sp+len(token)], token) {
+		if ascii.EqualFold(v[sp:sp+len(token)], token) {
 			return true
 		}
 	}
