@@ -6,10 +6,23 @@
 
 package http
 
-// RoundTrip implements the RoundTripper interface.
+import _ "unsafe" // for linkname
+
+// RoundTrip should be an internal detail,
+// but widely used packages access it using linkname.
+// Notable members of the hall of shame include:
+//   - github.com/erda-project/erda-infra
+//
+// Do not remove or change the type signature.
+// See go.dev/issue/67401.
+//
+//go:linkname badRoundTrip github.com/wangluozhe/chttp.(*Transport).RoundTrip
+func badRoundTrip(*Transport, *Request) (*Response, error)
+
+// RoundTrip implements the [RoundTripper] interface.
 //
 // For higher-level HTTP client support (such as handling of cookies
-// and redirects), see Get, Post, and the Client type.
+// and redirects), see [Get], [Post], and the [Client] type.
 //
 // Like the RoundTripper interface, the error types returned
 // by RoundTrip are unspecified.
